pages restricted

+++ base/views.py
@@ -1,19 +1,22 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from .models import Post, Topic
 from .forms import PostForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
 
-
-        
         try:
             user = User.objects.get(username=username)
         except:
@@ -55,7 +58,7 @@ def post(request, pk):
     context = {'post': post}
     return render(request, 'base/post.html', context)
 
-
+@login_required(login_url='login')
 def createPost(request):
     form = PostForm()
     if request.method == 'POST':
@@ -66,9 +69,14 @@ def createPost(request):
     context = {'form': form}
     return render(request, 'base/post_form.html', context)
 
+
+@login_required(login_url='login')
 def updatePost(request, pk):
     post = Post.objects.get(id=pk)
     form = PostForm(instance=post)
+
+    if request.user != post.creator:
+        return HttpResponse('You are not allowed here')
 
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
@@ -81,8 +89,13 @@ def updatePost(request, pk):
 
 
 
+@login_required(login_url='login')
 def deletePost(request, pk):
     post = Post.objects.get(id=pk)
+
+    if request.user != post.creator:
+        return HttpResponse('You are not allowed here')
+
     if request.method == "POST":
         post.delete()
         return redirect('home')
